refactor: RequestHandler 분리
