+++ tomcat/src/main/java/org/apache/coyote/http11/auth/AuthService.java
@@ -1,0 +1,5 @@
+package org.apache.coyote.http11.auth;
+
+public class AuthService {
+
+}
